Add SYCL error reporting and refactor headers
